requirements.txt
pyserial
smbus
RPi.bme280
